# Backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/addPatient": {
            "post": {
                "description": "Add a new patient by binding the request data and inserting into the repository",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Add a new patient",
                "parameters": [
                    {
                        "description": "User data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.ApplicationResponse"
                        }
                    },
                    "400": {
                        "description": "Failed to bind user data",
                        "schema": {
                            "$ref": "#/definitions/responses.ApplicationResponse"
                        }
                    },
                    "500": {
                        "description": "Error inserting user",
                        "schema": {
                            "$ref": "#/definitions/responses.ApplicationResponse"
                        }
                    }
                }
            }
        },
...
        "/users": {
            "get": {
                "description": "Fetch users by passing optional query parameters to filter results",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Retrieve users based on filters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by Addiction Type",
                        "name": "Addiction_Type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by Nasha Mukti Centre Code",
                        "name": "Nasha_Mukti_Centre_Code",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by Employment Status (integer)",
                        "name": "Employment_Status",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by Treatment Completion status (boolean)",
                        "name": "Is_Treatment_Completed",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of users",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, invalid query parameters",
                        "schema": {
                            "$ref": "#/definitions/responses.ApplicationResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized access",
                        "schema": {
                            "$ref": "#/definitions/responses.ApplicationResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ApplicationResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.User": {
            "type": "object",
            "properties": {
                "Addiction_Duration": {
                    "type": "integer"
                },
                "Addiction_Type": {
                    "type": "string"
                },
                "Age": {
                    "type": "integer"
                },
                "Counselling_Count": {
                    "type": "integer"
                },
                "Counsellor_Name": {
                    "type": "string"
                },
                "District": {
                    "type": "string"
                },
                "Duration_of-Treatment": {
                    "type": "integer"
                },
                "Employment_Status": {
                    "type": "integer"
                },
                "Gender": {
                    "type": "string"
                },
                "Guardian_Name": {
                    "type": "string"
                },
                "Is_Treatment_Completed": {
                    "type": "boolean"
                },
                "Joining_Date": {
                    "type": "string"
                },
                "Name": {
                    "type": "string"
                },
                "Nasha_Mukti_Centre_Address": {
                    "type": "string"
                },
                "Nasha_Mukti_Centre_Code": {
                    "type": "string"
                },
                "Nasha_Mukti_Centre_Name": {
                    "type": "string"
                },
                "State": {
                    "type": "string"
                },
                "Under_Treatment": {
                    "type": "boolean"
                },
                "_id": {
                    "type": "string"
                }
            }
        },
        "responses.ApplicationResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {}
            }
        }
    }
}